v0.2

+++ FanetSimSync/uavnet/applications/pingapp/PingTestApp.h
@@ -1,21 +1,3 @@
-//
-// Copyright (C) 2001, 2003, 2004 Johnny Lai, Monash University, Melbourne, Australia
-// Copyright (C) 2005 Andras Varga
-//
-// This program is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public License
-// as published by the Free Software Foundation; either version 2
-// of the License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with this program; if not, see <http://www.gnu.org/licenses/>.
-//
-
 #ifndef __INET_PINGAPP_H
 #define __INET_PINGAPP_H
 
@@ -46,11 +28,11 @@ class INET_API PingTestApp : public cSimpleModule, public ILifecycle
     L3Address destAddr;
     L3Address srcAddr;
     std::vector<L3Address> destAddresses;
-    int packetSize = 0;
+    double packetSize = 0;
     cPar *sendIntervalPar = nullptr;
     cPar *sleepDurationPar = nullptr;
-    int hopLimit = 0;
-    int count = 0;
+    double hopLimit = 0;
+    double count = 0;
     int destAddrIdx = -1;
     simtime_t startTime;
     simtime_t stopTime;
