+++ inetmanet-3.x/src/inet/linklayer/ieee80211/mac/channelaccess/Dcaf.cc
@@ -1,20 +1,3 @@
-//
-// Copyright (C) 2016 OpenSim Ltd.
-//
-// This program is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public License
-// as published by the Free Software Foundation; either version 2
-// of the License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with this program; if not, see http://www.gnu.org/licenses/.
-// 
-
 #include "Dcaf.h"
 #include "inet/common/ModuleAccess.h"
 #include "inet/linklayer/ieee80211/mac/contract/IRx.h"
@@ -44,7 +27,7 @@ void Dcaf::calculateTimingParameters()
 {
     slotTime = modeSet->getSlotTime();
     sifs = modeSet->getSifsTime();
-    int difs = par("difsTime");
+    double difs = par("difsTime");
     // The PIFS and DIFS are derived by the Equation (9-2) and Equation (9-3), as illustrated in Figure 9-14.
     // PIFS = aSIFSTime + aSlotTime (9-2)
     // DIFS = aSIFSTime + 2 × aSlotTime (9-3)
